feat: update wrapInTestContext to allow for usage with other Backends
