+++ packages/test-utils/package.json
@@ -23,6 +23,7 @@
     "gulp": "^4.0.2",
     "react": "^17.0.1",
     "react-dnd": "workspace:packages/react-dnd",
+    "react-dnd-html5-backend": "workspace:packages/backend-html5",
     "react-dnd-test-backend": "workspace:packages/backend-test",
     "react-dom": "^17.0.1"
   },
@@ -32,6 +33,7 @@
     "@types/react-dom": ">= 16",
     "react": ">= 16.14",
     "react-dnd": ">= 10.0.0",
+    "react-dnd-html5-backend": ">= 11.1.3",
     "react-dnd-test-backend": ">= 10.0.0",
     "react-dom": ">= 16.14"
   }
