+++ packages/test-utils/src/index.ts
@@ -1,1 +1,2 @@
+export * from './wrappers'
 export * from './utils'
